Add db store to keep active db sessions
To access a db instance you have to first create a new db or access an
existing one. If creation/access is successful the instance is added to
the store and you can access it. You can access an instance as long as
it is not expired, an instance expire after 5 min of inactivity
(expiration is configurable on store creation). After each access the
expiration is refreshed.